Use `rsassa-pss-sha256` instead of `rsassa-pss`.

+++ deploy/ci_setup.sh
@@ -6,7 +6,9 @@ docker rm --force ota_tuf-mariadb || true
 docker rm --force tuf-vault || true
 
 # Some jobs don't behave, nuke them all
-docker ps -q | xargs docker rm --force || true
+if [[ `docker ps -q | wc -l` -gt 0 ]]; then
+    docker ps -q | xargs docker rm --force || true
+fi
 
 mkdir ota_tuf_entrypoint.d/ || true
